Revert "Merge High Availabillity features into blockchain-service "

+++ migrations/202010171600125-update-handler-ids-add-timestamp.js
@@ -4,11 +4,11 @@ module.exports = {
             'handler_ids',
             'timestamp',
             {
-                type: Sequelize.BIGINT,
+                type: Sequelize.INTEGER,
             },
         );
         const now = Date.now();
-        await queryInterface.sequelize.query(`UPDATE handler_ids SET timestamp = ${now} `);
+        await queryInterface.sequelize.query(`UPDATE handler_ids SET timestamp = ${now}`);
     },
     down: async (queryInterface) => {
         await queryInterface.removeColumn('handler_ids', 'timestamp');
